Fix window bugs
- Fix SML "Pause" and "game over" window bugs due to not taking WX and
  WY properly into account

+++ graphics.h
@@ -12,10 +12,10 @@ extern int draw_scanline(s_emu *emu);
 extern int lockscreen(s_screen *scr);
 extern void ppu_modes_and_scanlines(s_emu *emu);
 extern void render_frame_and_vblank_if_needed(s_emu *emu);
-extern int draw_window(s_emu *emu, size_t i, uint8_t *pixel);
-extern int draw_background(s_emu *emu, size_t i, uint8_t *pixel);
-extern int draw_OBJ(s_emu *emu, size_t i, uint8_t *pixel, uint8_t sptd[SPRITES_PER_SCANLINE], uint8_t nb_sptd);
-extern int draw_OBJ_tile(s_emu *emu, size_t i, uint8_t *pixel, uint8_t sptd);
+extern int draw_window(s_emu *emu, int i, uint8_t *pixel);
+extern int draw_background(s_emu *emu, int i, uint8_t *pixel);
+extern int draw_OBJ(s_emu *emu, int i, uint8_t *pixel, uint8_t sptd[SPRITES_PER_SCANLINE], uint8_t nb_sptd);
+extern int draw_OBJ_tile(s_emu *emu, int i, uint8_t *pixel, uint8_t sptd);
 extern int DMA_transfer(s_emu *emu);
 
 #endif //GRAPHICS_H
